Added the ability to build native exe via GraalVM. Code clean up and refined initialize database command.

+++ app/src/main/kotlin/h2databasetool/cmd/InitializeDatabaseCommand.kt
@@ -4,15 +4,21 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.options.*
+import com.github.ajalt.clikt.parameters.types.choice
+import com.github.ajalt.mordant.rendering.TextStyles.bold
 import h2databasetool.BuildInfo
-import h2databasetool.cmd.ui.Style
+import h2databasetool.cmd.ui.Style.boldEmphasis
 import h2databasetool.cmd.ui.Style.h1
-import h2databasetool.env.Env
+import h2databasetool.cmd.ui.Style.notice
+import h2databasetool.cmd.ui.Style.softFocus
 import h2databasetool.commons.*
+import h2databasetool.commons.terminal.NL
 import h2databasetool.commons.terminal.echoMarkdown
 import h2databasetool.commons.terminal.fail
+import h2databasetool.env.Env
 import org.h2.jdbcx.JdbcDataSource
 import java.io.File
+import java.sql.SQLException
 import javax.sql.DataSource
 
 class InitializeDatabaseCommand : CliktCommand(NAME) {
@@ -27,30 +33,56 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
     )
 
     override fun help(context: Context): String {
-        val cmd = Style.boldEmphasis(BuildInfo.APP_EXE)
+        val cmd = boldEmphasis(BuildInfo.APP_EXE)
+
+        val forceInitChoices = {
+            buildString() {
+                Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.choices().onEachIndexed { i, (choice, mode) ->
+                    appendLine(
+                        "| ", (i + 1), ". ", bold(choice), bold(": "),
+                        when (mode) {
+                            Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.ALWAYS -> "Always attempt to delete the whole database."
+                            Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.FAIL -> "Fails with an error"
+                            Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.SKIP -> "Do not attempt to re-initialize the database"
+                            Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.INIT_ONLY_SCHEMAS -> """
+                                    |Only attempt to initialize schemas (if present),
+                                    | by first deleting the schema and then apply the
+                                    | supplied script.""".stripMultiLineToMargin()
+                        }
+                    )
+                }
+            }
+        }
         return """
-             Create a new database in the specified base directory.
-             
-             ${h1("EXAMPLES")}
-             
-            1. Create a coin collection
-               ```shell
-               $cmd initDb myCoinDb
-               ```
-            2. Create my business database with a customer and stock schemas.
-               ```shell
-               $cmd initDb myBizDb --init-schema customer --init-schema stock 
-               ```
-            3. Force the re-creation of schemas the business database
-               ```shell
-               $cmd initDb myBizDb --init-schema customer --init-schema stock --force
-               ```
-            4. Initialize a database named __finTackDb__ with a schema called expenses which needs to be set up
-               with a script `../templates/expenses-template.sql` 
-               ```shell
-               $cmd initDb finTrackDb --init-schema expenses ../template/expenses-template.sql
-               ```
-            """.trimIndent()
+            |Create a new database in the specified base directory.
+            |$NL
+            |Note that the forcing the re initializing of an existing database may result
+            |in a failure. This behaviour can be controlled via the `--force-init` option which provides 
+            |the following choices:
+            | 
+             ${forceInitChoices()}
+            | 
+            |${h1("EXAMPLES")}
+            | 
+            |1. Create a coin collection
+            |   ```shell
+            |   $cmd initDb myCoinDb
+            |   ```
+            |2. Create my business database with a customer and stock schemas.
+            |   ```shell
+            |   $cmd initDb myBizDb --init-schema customer --init-schema stock 
+            |   ```
+            |3. Force the re-creation of schemas the business database
+            |   ```shell
+            |   $cmd initDb myBizDb --init-schema customer --init-schema stock --force-init always
+            |   ```
+            |4. Initialize a database named __finTackDb__ with a schema called expenses which needs to be set up
+            |   with a script `../templates/expenses-template.sql` 
+            |   ```shell
+            |   $cmd initDb finTrackDb --init-schema expenses ../template/expenses-template.sql
+            |   ```
+            |""".trimMargin()
+
     }
 
     private val dataDir by option(metavar = "H2 data directory", envvar = Env.H2TOOL_DATA_DIR.envVariable)
@@ -61,9 +93,6 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
         .flag(default = false)
         .help("Do a dry run by printing command output to STDOUT.")
 
-    private val forceInitIfExists by option("--force")
-        .help("Overwrite existing database if exists")
-        .flag()
 
     private val quoted by option(
         "--quote-schema-name",
@@ -95,7 +124,7 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
         ).convert { it.file(canonical = true, absolute = true) }
 
 
-    private val schemaInitializers by option("--init-schema", "-s", metavar = "<schema [file]>")
+    private val initSchemas by option("--init-schema", "-s", metavar = "<schema [file]>")
         .help(
             """
             Create one or more schemas, and optionally execute a supplied sql script file against the schema.
@@ -103,11 +132,16 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
         ).varargValues(1, 2).transformAll() { optionArgs: List<List<String>> ->
             optionArgs.map { a ->
                 SchemaInitializer(
-                    schema = a.first(),
+                    schema = a.first().let { name -> if (quoted) "\"$name\"" else name },
                     schemaInitScript = a.secondOrNull()?.let(::File)
                 )
             }
         }.unique()
+
+    private val forceInit by option("--force-init")
+        .choice(Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.choices())
+        .default(Env.H2TOOL_DB_FORCE_INIT_MODE.default)
+        .help(Env.H2TOOL_DB_FORCE_INIT_MODE.description)
 
     private val database by argument(
         name = "database-name", help = "Name of the new database to initialize."
@@ -123,21 +157,39 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
             )
         }
 
-        val baseDir = dataDir
-            .file().apply {
-                if (!exists() && !mkdirs())
-                    fail("Unable to create base directory: $this")
+        val baseDir = dataDir.file().also { baseDir ->
+            if (!baseDir.exists()) doAction("Base data dir not found. Attempting to create : $baseDir") {
+                if (!baseDir.mkdirs())
+                    fail("Unable to create base directory: $baseDir")
             }
+        }
 
         val jdbcUrl = "jdbc:h2:$baseDir/$database"
+        var performInitDb = true
+        var performInitDbSchemas = true
 
         if (baseDir.containsDb(database)) {
-            if (!forceInitIfExists)
-                fail("Database $database already exists at $baseDir")
-            else doAction("Removing database $database from data directory: $baseDir") {
-                baseDir.files { (_, name, _) -> (name == database) }.forEach { file ->
-                    if (!file.delete())
-                        fail("Unable to delete data base file: $file")
+            when (forceInit) {
+
+                Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.ALWAYS -> doAction("Forcing re-initializing the database $jdbcUrl.") {
+                    baseDir.files { (_, name, _) -> (name == database) }.forEach { file ->
+                        if (!file.delete())
+                            fail("Unable to delete data base file: $file")
+                    }
+                }
+
+                Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.FAIL -> doAction("Not allowed to force initialize database $jdbcUrl.") {
+                    fail("Not allowed to force initializing database $jdbcUrl")
+                }
+
+                Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.SKIP -> doAction("Skip initializing database $jdbcUrl") {
+                    performInitDb = false
+                    performInitDbSchemas = false
+                }
+
+                Env.H2TOOL_DB_FORCE_INIT_MODE.Choice.INIT_ONLY_SCHEMAS -> doAction("Only init schemas in database $jdbcUrl") {
+                    performInitDbSchemas = true
+                    performInitDb = false
                 }
             }
         }
@@ -149,10 +201,10 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
             )
         }
 
-        doAction("Initialize data source:$jdbcUrl") {
+        if (performInitDb || performInitDbSchemas) doAction("Initialize data source:$jdbcUrl") {
             val ds = datasource(jdbcUrl)
-            ds.processSchemaInitializers()
-            ds.processInitializer()
+            ds.takeIf({ performInitDbSchemas })?.processSchemaInitializers()
+            ds.takeIf({ performInitDb })?.processInitializer()
         }
     }
 
@@ -173,13 +225,13 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
 
     private fun DataSource.processSchemaInitializers() {
 
-        if (schemaInitializers.isEmpty())
+        if (initSchemas.isEmpty())
             return
 
-        for ((schema, schemaScriptFile) in schemaInitializers) {
+        for ((schema, schemaScriptFile) in initSchemas) {
             echo(buildString {
-                append("Initializing ${Style.notice(schema)} schema")
-                schemaScriptFile?.also { append(" ", Style.softFocus("(using script $schemaScriptFile)")) }
+                append("Initializing ${notice(schema)} schema")
+                schemaScriptFile?.also { append(" ", softFocus("(using script $schemaScriptFile)")) }
                 append('.')
             })
             using(connection) { executeScript("drop schema $schema if exists cascade") }
@@ -206,7 +258,12 @@ class InitializeDatabaseCommand : CliktCommand(NAME) {
         ds.setURL(jdbcUrl)
         ds.setUser(user)
         ds.setPassword(password)
-        ds.takeIf { testOnce }?.connection?.close()
+        if (testOnce) ds.runCatching { connection.close() }.onFailure { e ->
+            when (e) {
+                is SQLException -> fail("Unable to connect to $jdbcUrl: ${e.message} [errorCode=${e.errorCode}]")
+                else -> throw e
+            }
+        }
     }
 
 }
